+++ tests/test_multiple_in_and_out_port_components.py
@@ -232,10 +232,10 @@ class TestHEX(TestCase):
 
             def temperature_regimes(self, carrier):
                 temperatures = []
-                if carrier == 829940433102452838:
+                if carrier == "c8a29d9a-a4c0-4331-ab0a-2e4ff52e838d":
                     temperatures = [70.0, 65.0, 60.0]  # producer out
 
-                if carrier == 8725433194681736500139:
+                if carrier == "8cc72543-f319-4d6e-8173-bd6ba500139b":
                     temperatures = [65.0, 60.0]  # first hex out
 
                 return temperatures
@@ -244,19 +244,16 @@ class TestHEX(TestCase):
                 constraints = super().constraints(ensemble_member)
 
                 carriers = self.temperature_carriers()
-                for carrier in carriers.values():
-                    carrier_map = carrier["id_number_mapping"]
-                    temperature_regimes = self.temperature_regimes(int(carrier_map))
+                for carrier_id in carriers.keys():
+                    temperature_regimes = self.temperature_regimes(carrier_id)
                     if len(temperature_regimes) > 1:
-                        carrier_var_name = str(carrier_map) + "_temperature"
+                        carrier_var_name = str(carrier_id) + "_temperature"
                         var_carrier = self.extra_variable(carrier_var_name, ensemble_member)
                         for i in range(var_carrier.shape[0] - 1):
                             constraints.append((var_carrier[i] - var_carrier[i + 1], 0.0, 0.0))
 
                         for temperature in temperature_regimes:
-                            selected_temp_vec = self.state_vector(
-                                f"{int(carrier_map)}_{temperature}"
-                            )
+                            selected_temp_vec = self.state_vector(f"{carrier_id}_{temperature}")
                             for i in range(var_carrier.shape[0] - 1):
                                 constraints.append(
                                     (selected_temp_vec[i] - selected_temp_vec[i + 1], 0.0, 0.0)
@@ -293,8 +290,10 @@ class TestHEX(TestCase):
         demand_matching_test(solution, results)
         energy_conservation_test(solution, results)
 
-        temp_prod = results["829940433102452838_temperature"]
-        temp_hex = results["8725433194681736500139_temperature"]
+        carrier_prod_id = "c8a29d9a-a4c0-4331-ab0a-2e4ff52e838d"
+        carrier_hex_id = "8cc72543-f319-4d6e-8173-bd6ba500139b"
+        temp_prod = results[f"{carrier_prod_id}_temperature"]
+        temp_hex = results[f"{carrier_hex_id}_temperature"]
 
         # check heat exchanger 1 is bypassed
         np.testing.assert_allclose(results[f"{hex_active_id}.__disabled"][:-1], 0)